+++ CBV_fb_clone/settings/base.py
@@ -39,7 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework_simplejwt',
-    #own
+    #own,
     'rest_framework',
     'register',
     'post',
